[update] save and deleteById

+++ models/product.js
@@ -1,6 +1,7 @@
 const fs = require('fs')
 const path = require('path')
 const root = require('../utils/path')
+const Cart = require('./cart')
 
 const p = path.join(root, 'data', 'products.json')
 
@@ -15,8 +16,8 @@ const getProductsFromFile = cb => {
 }
 
 module.exports = class Product {
-  constructor (title, imageUrl, price, description) {
-    this.id = Math.random().toString()
+  constructor (title, imageUrl, price, description, id) {
+    this.id = id
     this.title = title
     this.imageUrl = imageUrl
     this.price = price
@@ -25,9 +26,31 @@ module.exports = class Product {
 
   save () {
     getProductsFromFile(products => {
-      products.push(this)
-      fs.writeFile(p, JSON.stringify(products), err => {
-        console.log(err)
+      const existingProductIndex = products.findIndex(prod => prod.id === this.id)
+      if (existingProductIndex !== -1) {
+        const updatedProducts = [...products]
+        updatedProducts[existingProductIndex] = this
+        fs.writeFile(p, JSON.stringify(updatedProducts), err => {
+          console.log(err)
+        })
+      } else {
+        this.id = Math.random().toString()
+        products.push(this)
+        fs.writeFile(p, JSON.stringify(products), err => {
+          console.log(err)
+        })
+      }
+    })
+  }
+
+  static deleteById (id) {
+    getProductsFromFile(products => {
+      const product = products.find(prod => prod.id === id)
+      const updatedProducts = products.filter(p => p.id !== id)
+      fs.writeFile(p, JSON.stringify(updatedProducts), err => {
+        if (!err) {
+          Cart.deleteProduct(id, product.price)
+        }
       })
     })
   }
@@ -35,9 +58,6 @@ module.exports = class Product {
   static findById (id, cb) {
     getProductsFromFile(products => {
       const product = products.find(p => p.id === id)
-      console.log('')
-      console.log(product)
-      console.log('')
       cb(product)
     })
   }
